Tests (#3)
* added engine.ts tests

* added some lib test

* added coverage report

* added lib tests

* added typedArray tests

* fixed bug and jest config

* refactored tests paths

* changed compilation target

* fixed workflow

* new build

+++ src/typedArrays.ts
@@ -1,5 +1,5 @@
 export function packToArrayBuffer(data: [number, number][]) {
-    const buffer = new Float32Array(data.length * 2);
+    const buffer = new Float64Array(data.length * 2);
     for (let i = 0, j = 0; i < data.length; i++, j += 2) {
         buffer[j] = data[i][0];
         buffer[j + 1] = data[i][1];
@@ -8,7 +8,7 @@ export function packToArrayBuffer(data: [number, number][]) {
 }
 
 export function unpackFromArrayBuffer(arrayBuffer: ArrayBuffer) {
-    const buffer = new Float32Array(arrayBuffer);
+    const buffer = new Float64Array(arrayBuffer);
     const data: [number, number][] = [];
     for (let i = 0, j = 0; i < buffer.length; i += 2, j++) {
         data[j] = [buffer[i], buffer[i + 1]];
